+++ src/context/TodoContext.js
@@ -1,0 +1,54 @@
+import React from "react";
+
+import { useLocalStorage } from "../hooks/useLocalStorage";
+
+const TodoContext = React.createContext();
+
+function TodoProvider(children) {
+  const {
+    item: todos,
+    saveItem: saveTodos,
+    loading,
+    error,
+  } = useLocalStorage("TODOS_V1", []);
+  const [searchValue, setSearchValue] = React.useState("");
+
+  const totalTodos = todos.length;
+  const completedTodos = todos.filter((item) => !!item.completed).length;
+  const todosSearch = todos.filter((item) =>
+    item.text.toLocaleLowerCase().includes(searchValue.toLocaleLowerCase())
+  );
+
+  const completeTodo = (text) => {
+    const newTodos = [...todos];
+    const todoIndex = newTodos.findIndex((todo) => todo.text === text);
+    newTodos[todoIndex].completed = true;
+    saveTodos(newTodos);
+  };
+
+  const deleteTodo = (text) => {
+    const newTodos = [...todos];
+    const todoIndex = newTodos.findIndex((todo) => todo.text === text);
+    newTodos.splice(todoIndex, 1);
+    saveTodos(newTodos);
+  };
+  return (
+    <TodoContext.Provider
+      value={{
+        loading,
+        error,
+        totalTodos,
+        completedTodos,
+        searchValue,
+        setSearchValue,
+        todosSearch,
+        completeTodo,
+        deleteTodo,
+      }}
+    >
+      {children}
+    </TodoContext.Provider>
+  );
+}
+
+export { TodoContext, TodoProvider };
