+++ src/hooks/useLocalStorage.js
@@ -38,3 +38,41 @@ function useLocalStorage(itemName, initialValue) {
 }
 
 export { useLocalStorage };
+
+/* const defaultTodos = [
+    {
+      text: "Terminar curso de react",
+      completed: true,
+    },
+    {
+      text: "Ir a Platzi Conf",
+      completed: false,
+    },
+    {
+      text: "Pasear perrito",
+      completed: false,
+    },
+    {
+      text: "Llorar con la llorana bien mucho",
+      completed: false,
+    },
+    {
+      text: "Ir al gym",
+      completed: true,
+    },
+    {
+      text: "Lavar ropa",
+      completed: false,
+    },
+    {
+      text: "Terminar RE4",
+      completed: false,
+    },
+    {
+      text: "Tomar más cursos de JuanDC",
+      completed: false,
+    },
+  ]; */
+
+//localStorage.setItem('TODOS_V1', JSON.stringify(defaultTodos))
+//localStorage.removeItem('TODOS_V1')
